+++ src/input/archipelago/ArchipelagoManager.h
@@ -19,9 +19,9 @@ public:
     void Connect();
     bool IsValidSaveFile();
     void OnFrame();
-    void SetConfigPath(std::string configPath) const;
     bool IsWaitingForSaveFile();
     void OnSaveFileLoaded();
+    void SetServerConfiguration(const std::string& serverIP, const std::string& playerName, const std::string& serverPassword);
 
 private:
     void ManageMessages();
@@ -32,6 +32,11 @@ private:
     ConnectionStatus _status = ReadyForConnection;
 
 
+    std::string _serverIP;
+    std::string _serverPassword;
     std::string _playerName;
     std::string _seedName;
+
+    const float _suggestChangingConfigWaitTime = 2.5f;
+    std::clock_t _connectedAt;
 };