+++ src/mod.cpp
@@ -2,7 +2,7 @@
 
 #include "application/Randomizer.h"
 #include "input/archipelago/ArchipelagoManager.h"
-#include "input/autowin/AutoWinManager.h"
+#include "input/cheatsManager/CheatsManager.h"
 #include "input/eventDetector/EventDetector.h"
 #include "input/characterLoading/CharacterLoadingDetector.h"
 #include "output/archipelagoMessenger/ArchipelagoMessenger.h"
@@ -23,7 +23,7 @@ Randomizer randomizer = Randomizer(displayManager,
                                    checkRepository,
                                    archipelagoMessenger);
 
-AutoWinManager autoWinManager = AutoWinManager(randomizer);
+CheatsManager cheatsManager = CheatsManager(randomizer);
 ArchipelagoManager archipelagoManager = ArchipelagoManager(randomizer);
 EventDetector eventDetector = EventDetector(randomizer);
 CharacterLoadingDetector characterLoadingDetector = CharacterLoadingDetector(randomizer);
@@ -31,15 +31,31 @@ CharacterLoadingDetector characterLoadingDetector = CharacterLoadingDetector(ran
 
 __declspec(dllexport) void __cdecl Init(const char* path, const HelperFunctions& helperFunctions)
 {
-    // const IniFile* config = new IniFile(std::string(path) + "\\config.ini");
-    archipelagoManager.SetConfigPath(std::string(path) + "\\config.ini");
+
+
+    const IniFile* settingsIni = new IniFile(std::string(path) + "\\config.ini");
+
+    if (!settingsIni)
+        return randomizer.ShowStatusInformation("Invalid Settings INI");
+
+    const std::string serverIp = settingsIni->getString("AP", "IP");
+    const std::string playerName = settingsIni->getString("AP", "PlayerName");
+    const std::string serverPassword = settingsIni->getString("AP", "Password");
+    
+    archipelagoManager.SetServerConfiguration(serverIp, playerName, serverPassword);
+
+    const bool autoSkipCutscenes = settingsIni->getBool("GameSettings", "AutoSkipCutscenes", true);
+    const bool skipCredits  = settingsIni->getBool("GameSettings", "SkippeableCredits", true);
+    const bool winButtonEnabled  = settingsIni->getBool("GameSettings", "AutoWinButton", false);
+
+    cheatsManager.SetCheatsConfiguration(autoSkipCutscenes, skipCredits, winButtonEnabled);
 }
 
 __declspec(dllexport) void __cdecl OnFrame()
 {
     archipelagoManager.OnFrame();
     displayManager.OnFrame();
-    autoWinManager.OnPlayingFrame();
+    cheatsManager.OnPlayingFrame();
     if (Current_CharObj2 != nullptr)
     {
         eventDetector.OnPlayingFrame();