+++ src/output/menuManager/MenuManager.h
@@ -3,7 +3,6 @@
 class MenuManager
 {
 public:
-    MenuManager();
 
     void UnlockCharacterSelection(int eventAddress);
     void LockCharacterSelection(int eventAddress);