+++ CHANGELOG.md
@@ -2,16 +2,35 @@
 
 All notable changes to this project will be documented in this file.
 
+## [v0.1.0] - 2024-07-08
+
+### Changed
+
+- Re-enabled mission/trial options on the menu.
+- Last Story takes the player directly to the Perfect Chaos fight.
+- The demo cutscenes no longer give checks.
+- Unlocks are shown on the character select screen and pause menu.
+    - It only shows the characters that are playable in the randomizer.
+- Prevent Supersonic for being selected early.
+
+### Added
+
+- Added message to check your configs if the connection takes too long.
+- Cutscenes are skipped automatically.
+- Skipping credits is enabled.
+- Configuration for skipping scenes/credits and auto-win button.
+
 ## [v0.0.5] - 2024-07-07
 
 ### Added
 
-- Added level emblems as locations
-    - You can configure how many missions types for each character
-    - You can choose 0 (except for sonic)
-- Added emblem requirement for Perfect Chaos
-    - You can select the percentage of emblems needed to fight Perfect Chaos
-- Automated the release files generation
+- Added level emblems as locations.
+    - You can configure how many missions types for each character.
+    - You can choose 0 (except for sonic).
+- Added emblem requirement for Perfect Chaos.
+    - You can select the percentage of emblems needed to fight Perfect Chaos.
+- Added option to randomize or not the story unlocks.
+- Automated the release files generation.
 
 ## [v0.0.4] - 2024-07-04